Start adding ITensorNetwork code
